Extract common code into CompactionJobCompletionUtils
