+++ java/compaction/compaction-core/src/main/java/sleeper/compaction/job/completion/CompactionJobCompletion.java
@@ -20,12 +20,12 @@ import org.slf4j.LoggerFactory;
 
 import sleeper.compaction.job.CompactionJob;
 import sleeper.compaction.job.CompactionJobStatusStore;
-import sleeper.core.statestore.FileReference;
 import sleeper.core.statestore.StateStore;
 import sleeper.core.statestore.StateStoreException;
-import sleeper.core.statestore.exception.FileReferenceNotAssignedToJobException;
 import sleeper.core.util.ExponentialBackoffWithJitter;
 import sleeper.core.util.ExponentialBackoffWithJitter.WaitRange;
+
+import static sleeper.compaction.job.completion.CompactionJobCompletionUtils.updateStateStoreSuccess;
 
 public class CompactionJobCompletion {
     public static final Logger LOGGER = LoggerFactory.getLogger(CompactionJobCompletion.class);
@@ -55,37 +55,9 @@ public class CompactionJobCompletion {
 
     public void apply(CompactionJobCompletionRequest request) throws StateStoreException, InterruptedException {
         CompactionJob job = request.getJob();
-        updateStateStoreSuccess(job, request.getRecordsWritten());
+        updateStateStoreSuccess(job, request.getRecordsWritten(), stateStoreProvider.getByTableId(job.getTableId()),
+                jobAssignmentWaitAttempts, jobAssignmentWaitBackoff);
         statusStore.jobFinished(job, request.buildRecordsProcessedSummary(), request.getTaskId());
-    }
-
-    private void updateStateStoreSuccess(CompactionJob job, long recordsWritten) throws StateStoreException, InterruptedException {
-        StateStore stateStore = stateStoreProvider.getByTableId(job.getTableId());
-        FileReference fileReference = FileReference.builder()
-                .filename(job.getOutputFile())
-                .partitionId(job.getPartitionId())
-                .numberOfRecords(recordsWritten)
-                .countApproximate(false)
-                .onlyContainsDataForThisPartition(true)
-                .build();
-
-        // Compaction jobs are sent for execution before updating the state store to assign the input files to the job.
-        // Sometimes the compaction can finish before the job assignment is finished. We wait for the job assignment
-        // rather than immediately failing the job run.
-        FileReferenceNotAssignedToJobException failure = null;
-        for (int attempts = 0; attempts < jobAssignmentWaitAttempts; attempts++) {
-            jobAssignmentWaitBackoff.waitBeforeAttempt(attempts);
-            try {
-                stateStore.atomicallyReplaceFileReferencesWithNewOne(job.getId(), job.getPartitionId(), job.getInputFiles(), fileReference);
-                LOGGER.info("Atomically replaced {} file references in state store with file reference {}.", job.getInputFiles(), fileReference);
-                return;
-            } catch (FileReferenceNotAssignedToJobException e) {
-                LOGGER.warn("Job not yet assigned to input files on attempt {} of {}: {}",
-                        attempts + 1, jobAssignmentWaitAttempts, e.getMessage());
-                failure = e;
-            }
-        }
-        throw new TimedOutWaitingForFileAssignmentsException(failure);
     }
 
     @FunctionalInterface
