+++ java/system-test/system-test-dsl/pom.xml
@@ -117,6 +117,13 @@
         </dependency>
         <dependency>
             <groupId>sleeper</groupId>
+            <artifactId>compaction-job-execution</artifactId>
+            <version>${project.parent.version}</version>
+            <type>test-jar</type>
+            <scope>test</scope>
+        </dependency>
+        <dependency>
+            <groupId>sleeper</groupId>
             <artifactId>ingest-core</artifactId>
             <version>${project.parent.version}</version>
             <type>test-jar</type>
